Move LoadProofParameters to gtest/utils.cpp
