+++ src/gtest/test_mempoollimit.cpp
@@ -7,6 +7,7 @@
 
 #include "arith_uint256.h"
 #include "mempool_limit.h"
+#include "gtest/utils.h"
 #include "utiltime.h"
 #include "utiltest.h"
 #include "transaction_builder.h"
