+++ src/gtest/utils.h
@@ -1,7 +1,11 @@
 #ifndef BITCOIN_GTEST_UTILS_H
 #define BITCOIN_GTEST_UTILS_H
 
+#include "util.h"
+#include "key_io.h"
+
 int GenZero(int n);
 int GenMax(int n);
+void LoadProofParameters();
 
 #endif
